issue #17
The fix was to not check `bump` on the version object and instead use a different variable. When all checks pass, `bump` on the version object is set to the appropriate value. Tests have also been updated to deal with this case.

The larger issue is that the version number can change and is not finalized because we are using tasks. Need a better way to deal with this.